Index

InferenceReport.Inference

InferenceReport.PostprocessContext

InferenceReport.ReportOptions

InferenceReport.TargetDescription

InferenceReport.add_markdown

InferenceReport.add_plot

InferenceReport.add_table

InferenceReport.as_doc_page

InferenceReport.default_postprocessors

InferenceReport.gcb_progress

InferenceReport.get_pt

InferenceReport.headless

InferenceReport.intervals

InferenceReport.intervals

InferenceReport.lcb

InferenceReport.logz_progress

InferenceReport.moments

InferenceReport.moving_pair_plot

InferenceReport.mpi_standard_out

InferenceReport.pair_plot

InferenceReport.pigeons_inputs

InferenceReport.pigeons_summary

InferenceReport.pigeons_target_description

InferenceReport.pigeons_target_name

InferenceReport.render

InferenceReport.report

InferenceReport.report

InferenceReport.report_to_docs

InferenceReport.reproducibility_info

InferenceReport.round_trip_progress

InferenceReport.swaps_plot

InferenceReport.target_description

InferenceReport.target_name

InferenceReport.trace_plot

InferenceReport.trace_plot_cumulative

InferenceReport.view_webpage

InferenceReport.@reproducible

Types and functions

InferenceReport.Inference — Type

The object of a report created by this package:

algorithm: Algorithm that created the samples (currently assumes it is a PT struct). Optional, nothing if unspecfied.

chains: Samples, stored in a Chains struct.

original_dim: The original dimensionality of chains (the one stored might be truncated via options.max_dim)

InferenceReport.PostprocessContext — Type

Information passed to the processors:

inference: The Inference on which the report will be based on.

output_directory: The resolved location of the output report.

generated_markdown: Vector of Strings used internally to create the markdown page.

generated_dict: Generated dictionary used internally to create a JSON object for indexing purposes.

options: The ReportOptions.

InferenceReport.ReportOptions — Type

Options to control report:

max_moving_plot_iters: Maximum number of iterations in moving_pair_plot. Set to zero to avoid creating such movie.

Creating movies are a bit slow, so in interactive mode the default is zero, and 100 in headless mode.

target_name: Name of the target. If nothing, will attempt to auto-detect a name via multiple dispatch of the functions target_name or pigeons_target_name.

target_description: Description of the target, specified using a TargetDescription. If nothing, will attempt to auto-detect it via target_description and pigeons_target_description.

render: If true, the report webpage's md files are rendered into html files.

view: If true, the report webpage will be opened automatically.

postprocessors: The postprocessors to use. Default is default_postprocessors.

exec_folder: The directory where the report will be created. A unique folder in results/all will be created by default and symlinked in results/latest.

reproducibility_command: A command to instruct users how to reproduce.

At the moment implicitly assumes Pigeons will be used, and running these commands should return the Inputs.

writer: A Documenter.Writer instance used to render the document. Default is an HTML page, but other options including Documenter.LaTeX() are also possible.

show_error_traces: Useful for debugging, false by default.

max_dim: Upper bound on the number of dimensions to process. If the number of dimensions exceed this bound, only the prefix of dimensions will be processed.

interval_probability: Nominal coverage of intervals.

InferenceReport.TargetDescription — Type

Description of a target.

keywords: Keywords for the target.

bibliography: Contents of TeX bib file used to provide bibliographical references in the text, or nothing if no bibliography should be generated.

text: Description of the target.

InferenceReport.add_markdown — Method

add_markdown(context; title, contents)

InferenceReport.add_plot — Method

add_plot(context; file, title, url_help, description, movie)

InferenceReport.add_table — Method

add_table(
    context;
    table,
    title,
    url_help,
    description,
    kw_pretty_table_args...
)

InferenceReport.as_doc_page — Method

as_doc_page(context)

Pass to the pages argument of makedocs to properly include the generated page in the Documenter.jl navigation bar.

InferenceReport.default_postprocessors — Method

default_postprocessors()

InferenceReport.gcb_progress — Method

gcb_progress(context)

Progression of the global communication barrier estimate as a function of the round.

InferenceReport.get_pt — Method

get_pt(context)

InferenceReport.headless — Method

headless(lambda)

Enclose calls to report_to_docs and report with different defaults for the view and render options, setting them to false by default.

InferenceReport.intervals — Method

intervals(chain, probability)

InferenceReport.intervals — Method

intervals(context)

Table of credible intervals.

InferenceReport.lcb — Method

lcb(context)

The local communication barrier estimated in the last round.

InferenceReport.logz_progress — Method

logz_progress(context)

Progression of the log-normalization constant estimate as a function of the round.

InferenceReport.moments — Method

moments(context)

Table of posterior means, variances, MCSE, ESS, R-hat.

InferenceReport.moving_pair_plot — Method

moving_pair_plot(context)

A PairPlot movie showing the sample in the reference chain moving in the state space.

InferenceReport.mpi_standard_out — Method

mpi_standard_out(context)

MPI standard output.

InferenceReport.pair_plot — Method

pair_plot(context)

Create a pair plot using PairPlot.jl.

InferenceReport.pigeons_inputs — Method

pigeons_inputs(context)

Table of inputs used in Pigeons.

InferenceReport.pigeons_summary — Method

pigeons_summary(context)

The summary table produced by Pigeons during its execution.

InferenceReport.pigeons_target_description — Method

pigeons_target_description(target)

By default, dispatch to pigeons_target_description(target, Val(Symbol(pigeons_target_name(target)))), to handle "vague types" such as TuringLogPotential and StanLogPotential.

InferenceReport.pigeons_target_name — Method

pigeons_target_name(target)

InferenceReport.render — Method

render(context)

InferenceReport.report — Method

report(algo_or_chains, options)

InferenceReport.report — Method

report(algo_or_chains; args...)

algo_or_chains can be either a PT struct (returned by pigeons(..)) or a Chains.

args are used to create a ReportOptions.

InferenceReport.report_to_docs — Method

report_to_docs(algo_or_chains; doc_root, args...)

Similar to report but with different defaults to facilitate including the result as part of a larger documentation site.

Use in combination with as_doc_page.

InferenceReport.reproducibility_info — Method

reproducibility_info(context)

Generated reproducibility script.

InferenceReport.round_trip_progress — Method

round_trip_progress(context)

Progression of the number of round trips as a function of the round.

InferenceReport.swaps_plot — Method

swaps_plot(context)

A plot showing progression of mean swap acceptance per chain as a function of round.

InferenceReport.target_description — Method

target_description(context)

InferenceReport.target_name — Method

target_name(context)

InferenceReport.trace_plot — Method

trace_plot(context)

Trace plots.

InferenceReport.trace_plot_cumulative — Method

trace_plot_cumulative(context)

Cumulative averages as a function of iteration.

InferenceReport.view_webpage — Method

view_webpage(exec_folder)

InferenceReport.@reproducible — Macro

Returns a pair, where the first argument is the provided expression verbatim, and the second, the expression as a string to be used as the reproducibility_command argument in report and report_to_docs.