Index

InferenceReport.Inference

InferenceReport.PostprocessContext

InferenceReport.ReportOptions

InferenceReport.add_markdown

InferenceReport.add_plot

InferenceReport.add_table

InferenceReport.default_postprocessors

InferenceReport.gcb_progress

InferenceReport.logz_progress

InferenceReport.moments

InferenceReport.moving_pair_plot

InferenceReport.mpi_standard_out

InferenceReport.pair_plot

InferenceReport.pigeons_inputs

InferenceReport.pigeons_summary

InferenceReport.pigeons_target_name

InferenceReport.pigeons_target_name

InferenceReport.report

InferenceReport.report

InferenceReport.report_to_docs

InferenceReport.reproducibility_info

InferenceReport.round_trip_progress

InferenceReport.target_name

InferenceReport.target_name

InferenceReport.target_name

InferenceReport.target_name

InferenceReport.trace_plot

InferenceReport.trace_plot_cumulative

Types and functions

InferenceReport.Inference — Type

The object of a report created by this package:

algorithm: Detailed algorithm that created the samples (currently assumes it is a PT struct). Optional, nothing if unspecfied.

chains: Samples, stored in a Chains struct.

InferenceReport.PostprocessContext — Type

Information passed to the processors:

inference: The Inference on which the report will be based on.

output_directory: The resolved location of the output report.

generated_markdown: Vector of Strings used internally.

options: The ReportOptions.

InferenceReport.ReportOptions — Type

Options to control report:

max_moving_plot_iters: Maximum number of iterations in moving_pair_plot.

target_name: Name of the target. If nothing, target_name will be used to attempt to auto-detect it.

render: If true, the report webpage's md files are rendered into html files.

view: If true, the report webpage will be opened automatically.

postprocessors: The postprocessors to use. Default is default_postprocessors.

exec_folder: The directory where the report will be created. A unique folder in results/all will be created by default and symlinked in results/latest.

reproducibility_command: A command to instruct users how to reproduce.

At the moment implicitly assumes Pigeons will be used, and running these commands should return the Inputs.

InferenceReport.add_markdown — Method

add_markdown(context; title, contents)

InferenceReport.add_plot — Method

add_plot(context; file, title, description, movie)

InferenceReport.add_table — Method

add_table(context; table, title, kw_pretty_table_args...)

InferenceReport.default_postprocessors — Method

default_postprocessors()

InferenceReport.gcb_progress — Method

gcb_progress(context)

Progression of the global communication barrier estimate as a function of the round.

InferenceReport.logz_progress — Method

logz_progress(context)

Progression of the log-normalization constant estimate as a function of the round.

InferenceReport.moments — Method

moments(context)

Table of means and variances.

InferenceReport.moving_pair_plot — Method

moving_pair_plot(context)

A PairPlot movie showing the sample in the reference chain moving in the state space.

InferenceReport.mpi_standard_out — Method

mpi_standard_out(context)

MPI standard output.

InferenceReport.pair_plot — Method

pair_plot(context)

Create a pair plot using PairPlot.jl.

InferenceReport.pigeons_inputs — Method

pigeons_inputs(context)

Table of inputs used in Pigeons.

InferenceReport.pigeons_summary — Method

pigeons_summary(context)

The summary table produced by Pigeons during its execution.

InferenceReport.pigeons_target_name — Method

pigeons_target_name(target)

InferenceReport.pigeons_target_name — Method

pigeons_target_name(target)

InferenceReport.report — Method

report(algo_or_chains, options)

InferenceReport.report — Method

report(algo_or_chains; args...)

algo_or_chains can be either a PT struct (returned by pigeons(..)) or a Chains.

args are used to create a ReportOptions.

InferenceReport.report_to_docs — Method

report_to_docs(algo_or_chains; doc_root, args...)

Similar to report but as part of a larger documentation site.

Returns the correct relative path to the .md file to be passed to the pages argument of makedocs.

InferenceReport.reproducibility_info — Method

reproducibility_info(context)

Generated reproducibility script.

InferenceReport.round_trip_progress — Method

round_trip_progress(context)

Progression of the number of round trips as a function of the round.

InferenceReport.target_name — Method

target_name(unspecified_name, _)

InferenceReport.target_name — Method

target_name(unspecified_name, pt)

InferenceReport.target_name — Method

target_name(context)

InferenceReport.target_name — Method

target_name(specified_name, _)

InferenceReport.trace_plot — Method

trace_plot(context)

Trace plots.

InferenceReport.trace_plot_cumulative — Method

trace_plot_cumulative(context)

Cumulative averages as a function of iteration.